Refactor move directories
