+++ shared/src/commonMain/kotlin/org/hyperskill/app/users_questionnaire/questionnaire_onboarding/domain/analytic/QuestionnaireOnboardingAnalyticParams.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.app.questionnaire_onboarding.domain.analytic
+package org.hyperskill.app.users_questionnaire.questionnaire_onboarding.domain.analytic
 
 internal object QuestionnaireOnboardingAnalyticParams {
     const val PARAM_SOURCE = "source"
