+++ shared/src/commonMain/kotlin/org/hyperskill/app/users_questionnaire/questionnaire_onboarding/injection/QuestionnaireOnboardingComponentImpl.kt
@@ -1,9 +1,9 @@
-package org.hyperskill.app.questionnaire_onboarding.injection
+package org.hyperskill.app.users_questionnaire.questionnaire_onboarding.injection
 
 import org.hyperskill.app.core.injection.AppGraph
-import org.hyperskill.app.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.Action
-import org.hyperskill.app.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.Message
-import org.hyperskill.app.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.ViewState
+import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.Action
+import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.Message
+import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.ViewState
 import ru.nobird.app.presentation.redux.feature.Feature
 
 internal class QuestionnaireOnboardingComponentImpl(
