+++ shared/src/commonMain/kotlin/org/hyperskill/app/users_questionnaire/questionnaire_onboarding/presentation/QuestionnaireOnboardingActionDispatcher.kt
@@ -1,10 +1,10 @@
-package org.hyperskill.app.questionnaire_onboarding.presentation
+package org.hyperskill.app.users_questionnaire.questionnaire_onboarding.presentation
 
 import org.hyperskill.app.analytic.domain.interactor.AnalyticInteractor
 import org.hyperskill.app.core.presentation.ActionDispatcherOptions
-import org.hyperskill.app.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.Action
-import org.hyperskill.app.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.InternalAction
-import org.hyperskill.app.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.Message
+import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.Action
+import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.InternalAction
+import org.hyperskill.app.users_questionnaire.questionnaire_onboarding.presentation.QuestionnaireOnboardingFeature.Message
 import ru.nobird.app.presentation.redux.dispatcher.CoroutineActionDispatcher
 
 internal class QuestionnaireOnboardingActionDispatcher(
